project/backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
lightgbm==4.0.0
python-dotenv==1.0.0
